Set job timeout of 30m
The build and publish job can get stuck for various reasons and then need manual intervention. Set a timeout so they are canceled and retried in the next cron interval. 30m is very arbitrarily chosen, successful jobs only take a few minutes (or seconds if there is nothing to do).